+++ src/pkg.ts
@@ -1,6 +1,6 @@
 /** @format */
 
 export default {
-  version: '0.2.7-beta.1',
+  version: '0.2.7-beta.3',
   name: 'aliyun-pds-js-sdk',
 }